+++ Changelog.md
@@ -16,6 +16,7 @@
 - Added the `is_hidden` field to the `Property` model
 - Added the `withImages` method to the product search class (eager loads media)
 - Added the `channelables` table for being many-to-many polymorphic relationships with channels and arbitrary models
+- Added the `mapInto()` method to the `RelationAdjustmentCollection` class, which forwards the call to the underlying Eloquent collection
 
 ## 3.x Series